Clean up mixture_of_experts.py
